requirements.txt
asgiref==3.2.10
certifi==2019.9.11
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-crispy-forms==1.8.0
django-oauth-toolkit==1.2.0
djangorestframework==3.12.1
idna==2.8
oauthlib==3.1.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
